version 2.6
client:
* virtual navaid obs-style displays for instrument training
* display GPS status screen even if no GPS location received yet
* account for navaid dme antenna not same location as main navaid

server:
* support for virtual navaid (gs and dme info)
* server changes for update cycle ending 20160428
  BIE IAP-VOR RWY 36 CUEBE not correctly located
  EOD APD-AIRPORT DIAGRAM 87^28.5'W not correctly decoded
* iapreview.php checks missing plates as being listed as
  status 'D' on FAA webpage so don't have to manually verify
* update purgeoldcharts.sh

+++ decosects/purgeoldcharts.sh
@@ -22,6 +22,16 @@ function purgedatum
             rm -rf "$file"
             continue
         fi
+        if [ "${file:0:15}" == "datums/airways_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:18}" == "datums/apdgeorefs_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
         if [ "${file:0:19}" == "datums/aptdiags_150" ]
         then
             rm -rf "$file"
@@ -50,6 +60,31 @@ function purgedatum
             mv "$file" "purged-$file"
             continue
         fi
+        if [ "${file:0:12}" == "datums/fixes" ]
+        then
+            rm -rf "$file"
+            continue
+        fi
+        if [ "${file:0:20}" == "datums/getaptplates_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:18}" == "datums/iapgeorefs_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:20}" == "datums/intersections" ]
+        then
+            rm -rf "$file"
+            continue
+        fi
+        if [ "${file:0:17}" == "datums/localizers" ]
+        then
+            rm -rf "$file"
+            continue
+        fi
         if [ "${file:0:14}" == "datums/navaids" ]
         then
             rm -rf "$file"
@@ -58,6 +93,41 @@ function purgedatum
         if [ "${file:0:14}" == "datums/runways" ]
         then
             rm -rf "$file"
+            continue
+        fi
+        if [ "${file:0:16}" == "datums/waypoints" ]
+        then
+            rm -rf "$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/APT_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/AWY_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/FIX_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/ILS_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/NAV_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:11}" == "datums/TWR_" ]
+        then
+            mv "$file" "purged-$file"
             continue
         fi
     done
